add validation access method for update and delete

+++ src/main/java/com/cegeka/academy/web/rest/InvitationController.java
@@ -1,16 +1,20 @@
 package com.cegeka.academy.web.rest;
 
 import com.cegeka.academy.domain.Invitation;
-import com.cegeka.academy.repository.InvitationRepository;
 import com.cegeka.academy.service.invitation.InvitationService;
+import com.cegeka.academy.service.invitation.ValidationAccessService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
+
 
 @RestController
 @RequestMapping("invitation")
 public class InvitationController {
 
     private final InvitationService invitationService;
+
+    @Autowired
+    ValidationAccessService validationAccessService;
 
     @Autowired
     public InvitationController(InvitationService invitationService) {
@@ -29,11 +33,13 @@ public class InvitationController {
 
     @PutMapping("/update")
     public String replaceInvitation(@RequestBody Invitation newInvitation){
+        validationAccessService.verifyUserAccessForInvitationEntity(newInvitation.getId());
         return invitationService.updateInvitation(newInvitation);
     }
 
     @DeleteMapping("/invitation/{id}")
     void deleteInvitation(@PathVariable Long id) {
+        validationAccessService.verifyUserAccessForInvitationEntity(id);
         invitationService.deleteInvitationById(id);
     }
